move InstructionPtr into Wasmi executor
This is where it actually belongs.
