+++ crates/wasmi/src/engine/bytecode/mod.rs
@@ -1,5 +1,6 @@
 mod construct;
 mod immediate;
+mod instr_ptr;
 mod provider;
 mod utils;
 
@@ -8,6 +9,7 @@ mod tests;
 
 pub(crate) use self::{
     immediate::{AnyConst16, AnyConst32, Const16, Const32},
+    instr_ptr::InstructionPtr,
     provider::{Provider, ProviderSliceStack, UntypedProvider},
     utils::{
         BinInstr,
