+++ crates/wasmi/src/engine/code_map.rs
@@ -662,66 +662,3 @@ impl<'a> CompiledFuncRef<'a> {
         self.consts.get_ref()
     }
 }
-
-/// The instruction pointer to the instruction of a function on the call stack.
-#[derive(Debug, Copy, Clone)]
-pub struct InstructionPtr {
-    /// The pointer to the instruction.
-    ptr: *const Instruction,
-}
-
-/// It is safe to send an [`InstructionPtr`] to another thread.
-///
-/// The access to the pointed-to [`Instruction`] is read-only and
-/// [`Instruction`] itself is [`Send`].
-///
-/// However, it is not safe to share an [`InstructionPtr`] between threads
-/// due to their [`InstructionPtr::offset`] method which relinks the
-/// internal pointer and is not synchronized.
-unsafe impl Send for InstructionPtr {}
-
-impl InstructionPtr {
-    /// Creates a new [`InstructionPtr`] for `instr`.
-    #[inline]
-    pub fn new(ptr: *const Instruction) -> Self {
-        Self { ptr }
-    }
-
-    /// Offset the [`InstructionPtr`] by the given value.
-    ///
-    /// # Safety
-    ///
-    /// The caller is responsible for calling this method only with valid
-    /// offset values so that the [`InstructionPtr`] never points out of valid
-    /// bounds of the instructions of the same compiled Wasm function.
-    #[inline(always)]
-    pub fn offset(&mut self, by: isize) {
-        // SAFETY: Within Wasm bytecode execution we are guaranteed by
-        //         Wasm validation and Wasmi codegen to never run out
-        //         of valid bounds using this method.
-        self.ptr = unsafe { self.ptr.offset(by) };
-    }
-
-    #[inline(always)]
-    pub fn add(&mut self, delta: usize) {
-        // SAFETY: Within Wasm bytecode execution we are guaranteed by
-        //         Wasm validation and Wasmi codegen to never run out
-        //         of valid bounds using this method.
-        self.ptr = unsafe { self.ptr.add(delta) };
-    }
-
-    /// Returns a shared reference to the currently pointed at [`Instruction`].
-    ///
-    /// # Safety
-    ///
-    /// The caller is responsible for calling this method only when it is
-    /// guaranteed that the [`InstructionPtr`] is validly pointing inside
-    /// the boundaries of its associated compiled Wasm function.
-    #[inline(always)]
-    pub fn get(&self) -> &Instruction {
-        // SAFETY: Within Wasm bytecode execution we are guaranteed by
-        //         Wasm validation and Wasmi codegen to never run out
-        //         of valid bounds using this method.
-        unsafe { &*self.ptr }
-    }
-}
