+++ crates/wasmi/src/engine/executor/instrs.rs
@@ -15,12 +15,13 @@ use crate::{
             FuncIdx,
             GlobalIdx,
             Instruction,
+            InstructionPtr,
             Register,
             SignatureIdx,
             TableIdx,
             UnaryInstr,
         },
-        code_map::{CodeMap, InstructionPtr},
+        code_map::CodeMap,
         executor::stack::{CallFrame, FrameRegisters, ValueStack},
         DedupFuncType,
     },
